+++ src/utils/textEditorUtils.ts
@@ -1,22 +1,23 @@
 import { getConfigSetting } from '../config';
 import * as vscode from 'vscode';
 
-export async function showTextEditor(
-  uri: vscode.Uri | vscode.TextDocument,
-  preview = true
-): Promise<vscode.TextEditor> {
+export async function showTextEditor(options: {
+  uri: vscode.Uri | vscode.TextDocument;
+  viewColumn?: vscode.ViewColumn;
+  preview: boolean;
+}): Promise<vscode.TextEditor> {
   const config = getConfigSetting();
   let viewColumn = vscode.ViewColumn.Beside;
   if (config.responseViewColumn === 'current') {
     viewColumn = vscode.ViewColumn.Active;
   }
-  let document = uri;
+  let document = options.uri;
   if (document instanceof vscode.Uri) {
     document = await vscode.workspace.openTextDocument(document);
   }
   return await vscode.window.showTextDocument(document, {
-    viewColumn,
+    viewColumn: options.viewColumn || viewColumn,
     preserveFocus: config.responseViewPreserveFocus,
-    preview,
+    preview: options.preview,
   });
 }